SimpleFirstPersonController Mostly Done
Sprinting on left shift has been implemented: Sprinting works by having two variables for the walk speed and the sprint speed. Whenever shift is pressed it changes which speed is used to process movement.

Quits the application when escape is pressed

Crouching when left ctrl is pressed has been somewhat implemented: Crouching works by shrinking the play model by a crouchFactor decided in the inspector. Only issue is that this leads to the model following for a few seconds. Going to continue working on how to fix said bug.