Fixed Jump Bug along with general player oddities
Fixed Jump Bug: This was done by changing the collision layer of the player so that the ray cast wouldn't prematurely  hit it. Furthermore, the player property playerheight was renamed to ground detection length since that's what it actually represented. This number was then adjust to 3.33 since that seems to be the magic number which makes the beam long enough to just hit the ground.

Fixed player camera height: Discovered that the reason why the player's body could be seen earlier was due to the camera being too high. Lowered camera height then reenabled the mesh texture.

Sprinting change: Changed sprinting so that player can't change sprinting status while in air, if this change is disliked I can revert it next commit.