Created the pause menu and altered the main menu
Created a menu manager script to handle all menu elements. This includes pausing the game, resuming it, changing the volume, quitting the game, entering into the game, and returning the the main menu.

Replaced and deleted Play.cs in favor of the Menu Manager

Created a pause menu and options menu in which you can change the volume.

Made it so that the pause menu is accessed via the escape key.

Gave the options button and the quit button functionality on the main menu page based off the menu manager script